Moved `InvertSPD` method to seperate static `Utils` class and updated PCA models to use this
